Stop using `setup_requires`, add a `pyproject.toml`. (#45)
We have found that the use of `setup_requires` creates incompatible
builds between different minor versions of Python (because the
version-specific eggs of the setup requirements end up embedded in the
build output). `clr` was using `setup_requires` to install
`pytest-runner` which is itself deprecated. This commit deletes
`pytest-runner` and adds a `pyproject.toml` to trigger Pants'/Pex's
PEP-517/518 behavior.

I massaged how we specify dependencies & our docs to show the new way of
running tests after this change... not that anyone will be doing that
any time soon.